add an option to write to a CSV
This commit adds an option to write the PagerDuty user information to
a CSV file. It also implements `tabulate` to display the user data as
a nicely formatted table on stdout. Since we can write to a CSV now,
users can optionally use that in other streams for whatever they want
to make of this data.

Signed-off-by: jtimberman <joshua@symops.io>