update SSO script to return ARN
We're using the ARN for the OrgService identifier type, and we're not
using the identity store ID itself. This commit refactors the
IdentityStore class to use a named tuple for the elements returned
from the AWS ListInstances API call and use the `arn` and `store_id`
values for processing. It also renames the column in the output,
clarifies in the CSV writer the number of data (non-header) rows that
were written. Finally, we also remove the import for argparse, and we
ensure that if this is imported itself, we don't run `main()`.

Signed-off-by: Joshua Timberman <joshua@symops.io>